Move _function_wrapper up to Sampler class
